split up code, refactor
