+++ swap-orchestrator/src/compose.rs
@@ -352,46 +352,6 @@ pub trait IntoFlag {
     fn to_display(self) -> &'static str;
 }
 
-impl IntoFlag for monero::Network {
-    /// This is documented here:
-    /// https://docs.getmonero.org/interacting/monerod-reference/#pick-monero-network-blockchain
-    fn to_flag(self) -> Flag {
-        Flag(match self {
-            monero::Network::Mainnet => None,
-            monero::Network::Stagenet => Some("--stagenet".to_string()),
-            monero::Network::Testnet => Some("--testnet".to_string()),
-        })
-    }
-
-    fn to_display(self) -> &'static str {
-        match self {
-            monero::Network::Mainnet => "mainnet",
-            monero::Network::Stagenet => "stagenet",
-            monero::Network::Testnet => "testnet",
-        }
-    }
-}
-
-impl IntoFlag for bitcoin::Network {
-    /// This is documented here:
-    /// https://www.mankier.com/1/bitcoind
-    fn to_flag(self) -> Flag {
-        Flag(Some(match self {
-            bitcoin::Network::Bitcoin => "-chain=main".to_string(),
-            bitcoin::Network::Testnet => "-chain=test".to_string(),
-            _ => panic!("Only Mainnet and Testnet are supported"),
-        }))
-    }
-
-    fn to_display(self) -> &'static str {
-        match self {
-            bitcoin::Network::Bitcoin => "mainnet",
-            bitcoin::Network::Testnet => "testnet",
-            _ => panic!("Only Mainnet and Testnet are supported"),
-        }
-    }
-}
-
 pub trait IntoSpec {
     fn to_spec(self) -> String;
 }
@@ -424,7 +384,7 @@ impl IntoImageAttribute for OrchestratorImage {
 fn validate_compose(compose_str: &str) {
     serde_yaml::from_str::<Compose>(compose_str).unwrap_or_else(|_| {
         panic!(
-            "Generated compose spec to be valid. But it was not. This is the spec: \n\n{}",
+            "Expected generated compose spec to be valid. But it was not. This is the spec: \n\n{}",
             compose_str
         )
     });
