+++ swap-orchestrator/src/lib.rs
@@ -1,4 +1,6 @@
 pub mod asb;
+pub mod bitcoind;
 pub mod compose;
 pub mod electrs;
 pub mod images;
+pub mod monerod;
