+++ swap-orchestrator/src/main.rs
@@ -1,12 +1,14 @@
 mod asb;
+mod bitcoind;
 mod compose;
 mod electrs;
 mod images;
+mod monerod;
 mod prompt;
 
 use crate::compose::{
-    DOCKER_COMPOSE_FILE, IntoSpec, OrchestratorDirectories, OrchestratorImage, OrchestratorImages,
-    OrchestratorInput, OrchestratorNetworks,
+    IntoSpec, OrchestratorDirectories, OrchestratorImage, OrchestratorImages, OrchestratorInput,
+    OrchestratorNetworks, DOCKER_COMPOSE_FILE,
 };
 use std::path::PathBuf;
 use swap_env::config::{
